Slider feature tuning
The slider pauses the video whenever it is being moved by the mouse. This commit adds that after the slider has been moved, it will not just play but it will resume the state (paused or playing) of the video prior to moving the slider.

In short, if the video was playing before the slider gets moved, it will play after the slider gets moved. If the video was paused before the slider gets moved, it will play after the slider gets moved.