Code refactoring, volume functionality, and loop functionality
Lets start with the new functionality

1. The video now loops successfully (Yipee-Yipee-Yippe.-..)
2. The video has a functional volume slider (YIpe.)
3. The time slider now can update while it's still being dragged (it's finicky, but better then before) (ee-Yipee)

Now onto the changes that most of everyone will either love or hate me for.
I have refactored A LOT of the events code so that it is separated into functions (for primary controller only, secondary controller is messier then a Discord mod's room). The program works the same (minus the changes in this commit), the code is just in different spots. This is to help readability of the code and to find where things are. I suggest from here on out we use FXML events for simple events (like switching from primary to secondary controller). It is much easier to modify, bugfix, remove, and add functionality when the events are in code.